+++ src/test/java/org/folio/edge/connexion/MainVerticleSslAndTslTest.java
@@ -6,16 +6,20 @@ import io.vertx.core.file.FileSystemException;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.unit.TestContext;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
+import org.bouncycastle.jcajce.provider.BouncyCastleFipsProvider;
+import org.folio.edge.core.Constants;
 import org.folio.edge.core.utils.ApiKeyUtils;
 import org.folio.edge.core.utils.test.MockOkapi;
 import org.folio.edge.core.utils.test.TestUtils;
 import org.junit.After;
+import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 import org.junit.runner.RunWith;
 
 import java.io.IOException;
+import java.security.Security;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -24,11 +28,23 @@ import static org.folio.edge.core.Constants.SYS_REQUEST_TIMEOUT_MS;
 import static org.mockito.Mockito.spy;
 
 @RunWith(VertxUnitRunner.class)
-public class MainVerticleSslTest {
+public class MainVerticleSslAndTslTest {
+
+  private static final String KEYSTORE_TYPE = "BCFKS";
+  private static final String KEYSTORE_PATH = "test.keystore 1.bcfks";
+  private static final String TRUST_STORE_PATH = "test.truststore 1.bcfks";
+  private static final String KEYSTORE_PASSWORD = "SecretPassword";
+
   private Vertx vertx;
 
   @Rule
   public ExpectedException thrown = ExpectedException.none();
+
+  @Before
+  public void setUpOnce() {
+    Security.addProvider(new BouncyCastleFipsProvider());
+    vertx = Vertx.vertx();
+  }
 
   @After
   public void tearDownOnce() {
@@ -37,7 +53,7 @@ public class MainVerticleSslTest {
 
   @Test
   public void setupSslConfigWithoutType(TestContext context) throws Exception {
-    JsonObject config = getCommonConfig()
+    JsonObject config = getSslConfig()
         .put(SYS_HTTP_SERVER_SSL_ENABLED, true);
 
     thrown.expect(IllegalStateException.class);
@@ -48,7 +64,7 @@ public class MainVerticleSslTest {
 
   @Test
   public void setupSslConfigWithoutPath(TestContext context) throws Exception {
-    JsonObject config = getCommonConfig()
+    JsonObject config = getSslConfig()
         .put(SYS_HTTP_SERVER_SSL_ENABLED, true)
         .put(SYS_HTTP_SERVER_KEYSTORE_TYPE, "JKS");
 
@@ -60,7 +76,7 @@ public class MainVerticleSslTest {
 
   @Test
   public void setupSslConfigWithoutPassword(TestContext context) throws Exception {
-    JsonObject config = getCommonConfig()
+    JsonObject config = getSslConfig()
         .put(SYS_HTTP_SERVER_SSL_ENABLED, true)
         .put(SYS_HTTP_SERVER_KEYSTORE_TYPE, "JKS")
         .put(SYS_HTTP_SERVER_KEYSTORE_PATH, "sample_keystore.jks");
@@ -73,7 +89,7 @@ public class MainVerticleSslTest {
 
   @Test
   public void setupSslConfigWitInvalidPath(TestContext context) throws Exception {
-    JsonObject config = getCommonConfig()
+    JsonObject config = getSslConfig()
         .put(SYS_HTTP_SERVER_SSL_ENABLED, true)
         .put(SYS_HTTP_SERVER_KEYSTORE_TYPE, "JKS")
         .put(SYS_HTTP_SERVER_KEYSTORE_PATH, "some_keystore_path")
@@ -87,7 +103,7 @@ public class MainVerticleSslTest {
 
   @Test
   public void setupSslConfigWithNotValidPassword(TestContext context) throws Exception {
-    JsonObject config = getCommonConfig()
+    JsonObject config = getSslConfig()
         .put(SYS_HTTP_SERVER_SSL_ENABLED, true)
         .put(SYS_HTTP_SERVER_KEYSTORE_TYPE, "JKS")
         .put(SYS_HTTP_SERVER_KEYSTORE_PATH, "sample_keystore.jks")
@@ -101,7 +117,7 @@ public class MainVerticleSslTest {
 
   @Test
   public void setupCorrectSslConfig(TestContext context) throws Exception {
-    JsonObject config = getCommonConfig()
+    JsonObject config = getSslConfig()
         .put(SYS_HTTP_SERVER_SSL_ENABLED, true)
         .put(SYS_HTTP_SERVER_KEYSTORE_TYPE, "JKS")
         .put(SYS_HTTP_SERVER_KEYSTORE_PATH, "sample_keystore.jks")
@@ -127,7 +143,13 @@ public class MainVerticleSslTest {
     vertx.deployVerticle(MainVerticle.class.getName(), opt, context.asyncAssertSuccess());
   }
 
-  private JsonObject getCommonConfig() {
+  @Test
+  public void testSetupCorrectTslConfig(TestContext context) throws Exception {
+    JsonObject config = getTslConfig();
+    deployVerticle(context, config);
+  }
+
+  private JsonObject getSslConfig() {
     int serverPort = TestUtils.getPort();
     int okapiPort = TestUtils.getPort();
     return new JsonObject()
@@ -136,5 +158,18 @@ public class MainVerticleSslTest {
         .put(SYS_SECURE_STORE_PROP_FILE, "src/test/resources/ephemeral.properties")
         .put(SYS_LOG_LEVEL, "TRACE")
         .put(SYS_REQUEST_TIMEOUT_MS, 5000);
+  }
+
+  private JsonObject getTslConfig() {
+    int serverPort = TestUtils.getPort();
+    return new JsonObject().put(Constants.SYS_PORT, serverPort)
+        .put(Constants.SYS_HTTP_SERVER_SSL_ENABLED, true)
+        .put(Constants.SYS_HTTP_SERVER_KEYSTORE_TYPE, KEYSTORE_TYPE)
+        .put(Constants.SYS_HTTP_SERVER_KEYSTORE_PATH, KEYSTORE_PATH)
+        .put(Constants.SYS_HTTP_SERVER_KEYSTORE_PASSWORD, KEYSTORE_PASSWORD)
+        .put(Constants.SYS_WEB_CLIENT_SSL_ENABLED, true)
+        .put(Constants.SYS_WEB_CLIENT_TRUSTSTORE_TYPE, KEYSTORE_TYPE)
+        .put(Constants.SYS_WEB_CLIENT_TRUSTSTORE_PATH, TRUST_STORE_PATH)
+        .put(Constants.SYS_WEB_CLIENT_TRUSTSTORE_PASSWORD, KEYSTORE_PASSWORD);
   }
 }
